update: move camera permissions to processor
